+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/simple/ArrayTests.java
@@ -289,4 +289,122 @@ public class ArrayTests extends BaseSQLTests {
     public void testOutOfBounds() {
         this.runtimeConstantFail("SELECT ELEMENT(ARRAY [2, 3])", "array that does not have exactly 1 element");
     }
+
+    @Test
+    public void testArrayAppend() {
+        String ddl = "CREATE TABLE ARR_TBL(val INTEGER ARRAY NOT NULL)";
+
+        DBSPZSetLiteral input = new DBSPZSetLiteral(
+                new DBSPTupleExpression(
+                        new DBSPVecLiteral(
+                                new DBSPI32Literal(1),
+                                new DBSPI32Literal(2),
+                                new DBSPI32Literal(3))
+                ),
+                new DBSPTupleExpression(
+                        new DBSPVecLiteral(
+                                new DBSPI32Literal(5),
+                                new DBSPI32Literal(6),
+                                new DBSPI32Literal(7))
+                )
+        );
+        DBSPZSetLiteral result = new DBSPZSetLiteral(
+                new DBSPTupleExpression(
+                        new DBSPVecLiteral(
+                                new DBSPI32Literal(1),
+                                new DBSPI32Literal(2),
+                                new DBSPI32Literal(3),
+                                new DBSPI32Literal(4))
+                ),
+                new DBSPTupleExpression(
+                        new DBSPVecLiteral(
+                                new DBSPI32Literal(5),
+                                new DBSPI32Literal(6),
+                                new DBSPI32Literal(7),
+                                new DBSPI32Literal(4))
+                )
+        );
+
+        this.testQuery(ddl, "SELECT ARRAY_APPEND(val, 4) FROM ARR_TBL", new InputOutputPair(input, result));
+    }
+
+    @Test
+    public void testArrayAppendNullable() {
+        String ddl = "CREATE TABLE ARR_TBL(val INTEGER ARRAY)";
+
+        DBSPZSetLiteral input = new DBSPZSetLiteral(
+                new DBSPTupleExpression(
+                        new DBSPVecLiteral(true,
+                                new DBSPI32Literal(1),
+                                new DBSPI32Literal(2),
+                                new DBSPI32Literal(3))
+                ),
+                new DBSPTupleExpression(
+                        new DBSPVecLiteral(true,
+                                new DBSPI32Literal(5),
+                                new DBSPI32Literal(6),
+                                new DBSPI32Literal(7))
+                )
+        );
+        DBSPZSetLiteral result = new DBSPZSetLiteral(
+                new DBSPTupleExpression(
+                        new DBSPVecLiteral(true,
+                                new DBSPI32Literal(1),
+                                new DBSPI32Literal(2),
+                                new DBSPI32Literal(3),
+                                new DBSPI32Literal(4))
+                ),
+                new DBSPTupleExpression(
+                        new DBSPVecLiteral(true,
+                                new DBSPI32Literal(5),
+                                new DBSPI32Literal(6),
+                                new DBSPI32Literal(7),
+                                new DBSPI32Literal(4))
+                )
+        );
+
+        this.testQuery(ddl, "SELECT ARRAY_APPEND(val, 4) FROM ARR_TBL", new InputOutputPair(input, result));
+    }
+
+    @Test @Ignore("https://issues.apache.org/jira/browse/CALCITE-6275")
+    public void testArrayAppendInnerNullable() {
+        String ddl = "CREATE TABLE ARR_TBL(val INTEGER ARRAY NULL)";
+
+        DBSPZSetLiteral input = new DBSPZSetLiteral(
+                new DBSPTupleExpression(
+                        new DBSPVecLiteral(true,
+                                DBSPLiteral.none(new DBSPTypeInteger(CalciteObject.EMPTY, 32, true,true)),
+                                new DBSPI32Literal(1, true),
+                                new DBSPI32Literal(2, true),
+                                new DBSPI32Literal(3, true))
+                ),
+                new DBSPTupleExpression(
+                        new DBSPVecLiteral(true,
+                                DBSPLiteral.none(new DBSPTypeInteger(CalciteObject.EMPTY, 32, true,true)),
+                                new DBSPI32Literal(5, true),
+                                new DBSPI32Literal(6, true),
+                                new DBSPI32Literal(7, true))
+                )
+        );
+        DBSPZSetLiteral result = new DBSPZSetLiteral(
+                new DBSPTupleExpression(
+                        new DBSPVecLiteral(true,
+                                DBSPLiteral.none(new DBSPTypeInteger(CalciteObject.EMPTY, 32, true,true)),
+                                new DBSPI32Literal(1, true),
+                                new DBSPI32Literal(2, true),
+                                new DBSPI32Literal(3, true),
+                                new DBSPI32Literal(4, true))
+                ),
+                new DBSPTupleExpression(
+                        new DBSPVecLiteral(true,
+                                DBSPLiteral.none(new DBSPTypeInteger(CalciteObject.EMPTY, 32, true,true)),
+                                new DBSPI32Literal(5, true),
+                                new DBSPI32Literal(6, true),
+                                new DBSPI32Literal(7, true),
+                                new DBSPI32Literal(4, true))
+                )
+        );
+
+        this.testQuery(ddl, "SELECT ARRAY_APPEND(val, 4) FROM ARR_TBL", new InputOutputPair(input, result));
+    }
 }