+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#1384](https://github.com/feldera/feldera/pull/1384))
 - SQL: Allow tables definitions to be retrieved from JDBC data sources
   ([#1416](https://github.com/feldera/feldera/pull/1416))
+- SQL: Support for EXP function
+  ([#1406](https://github.com/feldera/feldera/pull/1406))
+- SQL: Support for ARRAY_APPEND function
+  ([#1419](https://github.com/feldera/feldera/pull/1419))
 - SQL: support for TUMBLE streaming windows
   ([#1404](https://github.com/feldera/feldera/pull/1404))
 - API: endpoint to get a list of demo URLs